requirements.txt
aiofiles==23.2.1
altair==5.2.0
anyio==4.2.0
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
contourpy==1.1.1
cycler==0.12.1
easyocr==1.7.1
etelemetry==0.3.1
exceptiongroup==1.2.0
filelock==3.13.1
fitz==0.0.1.dev2
fonttools==4.47.2
frontend==0.0.3
fsspec==2023.12.2
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.20.2
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
looseversion==1.3.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
nibabel==5.2.0
ninja==1.11.1.1
nipype==1.8.6
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.0.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==10.2.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
protobuf==4.25.2
prov==2.0.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pydeck==0.8.1b0
pydot==2.0.0
PyDrive==1.3.1
Pygments==2.17.2
PyMuPDF==1.18.19
pyparsing==3.1.1
PyPDF2==3.0.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytils==0.4.1
pytz==2023.3.post1
PyWavelets==1.4.1
pyxnat==1.6
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.35.1
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tifffile==2023.7.10
timm==0.9.12
tokenizers==0.15.0
toml==0.10.2
tools==0.1.9
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traits==6.3.2
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.26.0
validators==0.22.0
Wand==0.6.13
watchdog==3.0.0
zipp==3.17.0
